refactor(filteres): filters and states in map-container
