--- src/discrete-layer/views/discrete-layer-view.css
@@ -1,14 +1,0 @@
-.drawerPosition{
-  overflow: hidden; 
-  position: absolute;
-  top: 60px;
-  left: 0px;
-}
-
-.drawerPosition .mdc-drawer{
-  width: 100%;
-} 
-body[dir='rtl'] .drawerPosition {
-  right: 0px;
-  left: unset;
-}
